IMP - Refactor array form file
